+++ dto/src/main/java/dto/AirportDto.java
@@ -1,4 +1,4 @@
-package DTO;
+package dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -9,10 +9,15 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 
-public class CityDto {
+public class AirportDto {
+    @JsonProperty(value = "city_code")
+    private String cityCode;
     @JsonProperty(value = "country_code")
     private String countryCode;
     private String code;
     private String name;
-}
 
+    private boolean flightable;
+    @JsonProperty(value = "iata_type")
+    private String iataType;
+}
