+++ dto/src/main/java/dto/CityDto.java
@@ -1,4 +1,4 @@
-package DTO;
+package dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -9,15 +9,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 
-public class AirportDto {
-    @JsonProperty(value = "city_code")
-    private String cityCode;
+public class CityDto {
     @JsonProperty(value = "country_code")
     private String countryCode;
     private String code;
     private String name;
-
-    private boolean flightable;
-    @JsonProperty(value = "iata_type")
-    private String iataType;
 }
+
