+++ dto/src/main/java/dto/CountryDto.java
@@ -1,4 +1,4 @@
-package DTO;
+package dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
