+++ dto/src/main/java/dto/SearchRequestDto.java
@@ -1,4 +1,4 @@
-package DTO;
+package dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
