+++ dto/src/main/java/dto/SearchResultDto.java
@@ -1,4 +1,4 @@
-package DTO;
+package dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
