+++ dto/src/main/java/dto/SearchResultDtoList.java
@@ -1,4 +1,4 @@
-package DTO;
+package dto;
 
 import lombok.*;
