+++ searcher/src/main/java/ru/otus/searcher/controllers/AirportSearchController.java
@@ -1,7 +1,7 @@
 package ru.otus.searcher.controllers;
 
 
-import DTO.AirportDto;
+import dto.AirportDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
