+++ searcher/src/main/java/ru/otus/searcher/controllers/CitiesSearchController.java
@@ -1,7 +1,7 @@
 package ru.otus.searcher.controllers;
 
 
-import DTO.CityDto;
+import dto.CityDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
