+++ searcher/src/main/java/ru/otus/searcher/converter/TicketSearchResultToSearchResultDTOConverter.java
@@ -2,8 +2,8 @@ package ru.otus.searcher.converter;
 
 import ru.otus.searcher.model.Ticket;
 import ru.otus.searcher.model.TicketSearchResult;
-import DTO.SearchResultDto;
-import DTO.SearchResultDtoList;
+import dto.SearchResultDto;
+import dto.SearchResultDtoList;
 import lombok.experimental.UtilityClass;
 
 import java.util.*;
