--- searcher/src/main/java/ru/otus/searcher/model/CountryModel.java
@@ -1,12 +1,0 @@
-package com.example.searcher.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class CountryModel {
-    private String country_code;
-    private String code;
-    private String name;
-}
