+++ searcher/src/main/java/ru/otus/searcher/service/TicketListService.java
@@ -1,7 +1,7 @@
 package ru.otus.searcher.service;
 
-import DTO.SearchRequestDto;
-import DTO.SearchResultDtoList;
+import dto.SearchRequestDto;
+import dto.SearchResultDtoList;
 
 public interface TicketListService {
     SearchResultDtoList getDtoTicketList(SearchRequestDto dto);
