+++ searcher/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 travelpayout.token = ${TOKEN:bdd5ddc926ca9d51e9a35700060d1fd0}
 travelpayout.url = ${TRAVEL_PAYOUT_URL:api.travelpayouts.com}
 travelpayout.search-url= /v2/prices/latest
-travelpayout.countries-url= /aviasales_resources/v3/countries.json
-travelpayout.cities-url = /aviasales_resources/v3/cities.json
-travelpayout.airports-url = /aviasales_resources/v3/airports.json
+travelpayout.countries-path= /aviasales_resources/v3/countries.json
+travelpayout.cities-path = /aviasales_resources/v3/cities.json
+travelpayout.airports-path = /aviasales_resources/v3/airports.json
