+++ searcher/src/test/java/ru/otus/searcher/controllers/CountrySearchControllerTest.java
@@ -2,7 +2,7 @@ package ru.otus.searcher.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import DTO.CountryDto;
+import dto.CountryDto;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
